Modification du traitement des actions du joueur

+++ src/main/Main.java
@@ -18,11 +18,12 @@ public class Main extends ListenerAdapter {
 //	private boolean playing;
 	private static Map<String, Player> players = new HashMap<>();
 	private static JDA bot;
-	private final static String token = "NjU3MzQ1MzQ4MTQxODQyNDMy.Xf1coQ.zsdMGtOGMsdD8U-byXHt6i1gNgI";
+	private final static String token = "NjU3MzQ1MzQ4MTQxODQyNDMy.Xf1uGA.yJX4nW4CAlM4Kun7__PpSzk1AoA";
+	private static boolean clearing = false;
 
 	public static void main(String[] args) throws LoginException {
-		bot = new JDABuilder(token).setActivity(Activity.playing("!help pour plus d'infos")).addEventListeners(new Main())
-				.build();
+		bot = new JDABuilder(token).setActivity(Activity.playing("!help pour plus d'infos"))
+				.addEventListeners(new Main()).build();
 
 	}
 
@@ -32,7 +33,7 @@ public class Main extends ListenerAdapter {
 			String msg = event.getMessage().getContentDisplay();
 			switch (msg) {
 			case ("!play"): {
-				if (!players.containsKey(event.getAuthor().getId())) {
+				if (!players.containsKey(event.getAuthor().getId()) && !clearing) {
 					players.put(event.getAuthor().getId(), new Player(event.getAuthor()));
 					Game g = new Game(players.get(event.getAuthor().getId()), event.getChannel());
 					bot.addEventListener(g);
@@ -42,16 +43,21 @@ public class Main extends ListenerAdapter {
 				break;
 			}
 			case ("!clear"): {
-				for (Message iter : event.getChannel().getIterableHistory()) {
-					iter.delete().queue();
+				if (players.isEmpty()) {
+					clearing = true;
+					for (Message iter : event.getChannel().getIterableHistory()) {
+						iter.delete().queue();
+					}
 				}
+				clearing = false;
 				break;
 			}
 			case ("!help"): {
-				event.getChannel().sendMessage("Voici les commandes disponnible pour le moment "
+				event.getChannel().sendMessage("Voici les commandes disponible pour le moment "
 						+ event.getAuthor().getAsMention()
-						+ "\n!clear - Supprime tout les messages dans le channel\n!play - Lance une partie de Whack a mole\n--> Lors d'une partie de whack a mole vous pourrez gagner des point en tapant sur les taupe pour cela taper dans le chat le numero correspondant a la position de la taupe.\n Chaque taupe taper vous donnera 15 point.")
-						.queue();
+						+ "\n!clear - Supprime tout les messages dans le channel\n!play - Lance une partie de Whack a mole\n--> Lors d'une partie de whack a mole vous pourrez gagner des point en tapant sur les taupe pour cela taper dans le chat le numero correspondant a la position de la taupe.\n Chaque taupe taper vous donnera 15 points."
+						+ "--> Si la taupe n'est pas sortie du trou et que vous la tapez vous perdez 5 points"
+						+ "\n!stop - arrete la partie en cours").queue();
 			}
 			}
 			System.out.println(event.getAuthor());
